+++ blockfrost-api/CHANGELOG.md
@@ -4,6 +4,8 @@
 * Additions
   * `/accounts/:stake_address/addresses/total` endpoint and `AddressAssociatedTotal` type
   * `/pools/extended` endpoint and `Pool` type
+  * `/utils` API
+    * `/utils/addresses/xpub/:xpub/:role/:index` endpoint and `DerivedAddress` type
 
 # Version [0.8.1.0](https://github.com/blockfrost/blockfrost-haskell/compare/api-0.8.0.0...api-0.8.1.0) (2023-09-18)